Move RegistryCenter
